Move counsel_leaderboards to backend/services
Service-layer module belongs under backend/services/ alongside other
service modules. Pure relocation with no behavior change.
