+++ backend/routes/agreements.py
@@ -13,7 +13,7 @@ from flask_smorest import Blueprint
 from sqlalchemy import and_, asc, bindparam, desc, func, or_, text
 from sqlalchemy.exc import SQLAlchemyError
 
-from backend.counsel_leaderboards import build_counsel_leaderboards_from_summary_rows
+from backend.services.counsel_leaderboards import build_counsel_leaderboards_from_summary_rows
 from backend.filtering import (
     build_canonical_counsel_agreement_uuid_subquery,
     build_transaction_price_bucket_filter,
